Refactor to clean the main.rs and remove warnings from compiler
