+++ src/states/mod.rs
@@ -1,1 +1,2 @@
+pub mod main_game;
 pub mod paused;
